+++ rollup.config.js
@@ -1,20 +1,29 @@
 import typescript from '@rollup/plugin-typescript';
 import copy from 'rollup-plugin-cpy';
 
+
+const plugins = [
+  typescript(),
+  copy({
+    files: 'src/peregrine/*',
+    dest: 'dist/',
+    options: {
+      verbose: true
+    }
+  })
+];
+
 export default {
   input: 'src/index.ts',
-  output: {
-    dir: 'dist',
-    format: 'es'
-  },
-  plugins: [
-    typescript(),
-    copy({
-      files: 'src/peregrine/*.wasm*',
-      dest: 'dist',
-      options: {
-        verbose: true
-      }
-    })
-  ]
+  output: [
+    {
+      file: 'dist/bundle-cjs.js',
+      format: 'cjs'
+    },
+    {
+      file: 'dist/bundle-es.js',
+      format: 'es'
+    }
+  ],
+  plugins: plugins
 };
