--- src/peregrine/peregrine_ensembl_bg.wasm.d.ts
@@ -1,26 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-export const memory: WebAssembly.Memory;
-export function __wbg_genomebrowser_free(a: number): void;
-export function genomebrowser_new(): number;
-export function genomebrowser_go(a: number): void;
-export function genomebrowser_set_stick(a: number, b: number, c: number): void;
-export function genomebrowser_receive_message(a: number): void;
-export function genomebrowser_set_bp_per_screen(a: number, b: number): void;
-export function genomebrowser_set_x(a: number, b: number): void;
-export function genomebrowser_set_y(a: number, b: number): void;
-export function genomebrowser_set_switch(a: number, b: number): void;
-export function genomebrowser_clear_switch(a: number, b: number): void;
-export function genomebrowser_set_message_reporter(a: number, b: number): void;
-export function init_panic_hook(): void;
-export function main(): void;
-export function __wbindgen_malloc(a: number): number;
-export function __wbindgen_realloc(a: number, b: number, c: number): number;
-export const __wbindgen_export_2: WebAssembly.Table;
-export function _dyn_core__ops__function__Fn_____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h104d4ffbf6f999ea(a: number, b: number): void;
-export function _dyn_core__ops__function__Fn__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__hbdd24a3fc63c8719(a: number, b: number, c: number): void;
-export function _dyn_core__ops__function__Fn__A_B___Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__hfd1189ba57b7e2df(a: number, b: number, c: number, d: number): void;
-export function _dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h0dafd344521636c4(a: number, b: number, c: number): void;
-export function __wbindgen_free(a: number, b: number): void;
-export function __wbindgen_exn_store(a: number): void;
-export function __wbindgen_start(): void;
