+++ src/peregrine/peregrine_generic_bg.wasm.d.ts
@@ -1,0 +1,33 @@
+/* tslint:disable */
+/* eslint-disable */
+export const memory: WebAssembly.Memory;
+export function test(a: number): number;
+export function __wbg_genomebrowser_free(a: number): void;
+export function genomebrowser_new(): number;
+export function genomebrowser_go(a: number, b: number): void;
+export function genomebrowser_copy(a: number): number;
+export function genomebrowser_set_stick(a: number, b: number, c: number): void;
+export function genomebrowser_jump(a: number, b: number, c: number): void;
+export function genomebrowser_wait(a: number): void;
+export function genomebrowser_receive_message(a: number): void;
+export function genomebrowser_set_artificial(a: number, b: number, c: number, d: number): void;
+export function genomebrowser_goto(a: number, b: number, c: number): void;
+export function genomebrowser_set_y(a: number, b: number): void;
+export function genomebrowser_set_switch(a: number, b: number): void;
+export function genomebrowser_clear_switch(a: number, b: number): void;
+export function genomebrowser_radio_switch(a: number, b: number, c: number): void;
+export function genomebrowser_set_message_reporter(a: number, b: number): void;
+export function main(): void;
+export function init_panic_hook(): void;
+export function __wbindgen_malloc(a: number): number;
+export function __wbindgen_realloc(a: number, b: number, c: number): number;
+export const __wbindgen_export_2: WebAssembly.Table;
+export function _dyn_core__ops__function__Fn_____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h2288f0c0e7f37955(a: number, b: number): void;
+export function _dyn_core__ops__function__FnMut_____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h3eaee146da3697d3(a: number, b: number): void;
+export function _dyn_core__ops__function__Fn__A_B___Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h15b65a97306dce86(a: number, b: number, c: number, d: number): void;
+export function _dyn_core__ops__function__Fn__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__hcce910934e589a54(a: number, b: number, c: number): void;
+export function _dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h307b7650279a1344(a: number, b: number, c: number): void;
+export function __wbindgen_free(a: number, b: number): void;
+export function __wbindgen_exn_store(a: number): void;
+export function wasm_bindgen__convert__closures__invoke2_mut__he5c7a1f76df3ecdb(a: number, b: number, c: number, d: number): void;
+export function __wbindgen_start(): void;
